# chronos_app/packages/api-documentation/src/yml/agent-protocol.openapi.yaml
openapi: 3.1.0
info:
    title: Chronos Agent Protocol
    version: 1.6.0
    summary: Wire-format contract for the Chronos Agent Registry and MCP Gateway.
    description: |
        Two-direction protocol between the Chronos control plane and external HTTP
        agents.

        - **Platform → Agent.** What an external HTTP agent must implement to be
          invokable via the Chronos Agent Registry. Health probe + an OpenAI-compatible
          chat completions endpoint with two Chronos extension fields.
        - **Agent → Platform.** What Chronos exposes for tool brokering. Agents call
          back into the MCP Gateway to invoke registered MCP tools without holding
          MCP-server credentials directly.

        See `docs/agent-registry.md` and `docs/mcp-gateway.md` for the operator
        guides. v1.6 supports the `bearer` and `header` outbound-auth shapes;
        OAuth2 client-credentials is a follow-up. v1.6 supports `streamable-http`
        and `sse` MCP transports; `stdio` is reserved for v1.8.
    license:
        name: Apache 2.0
        identifier: Apache-2.0
        url: https://github.com/popularowl/chronos/blob/main/LICENSE.md
    contact:
        name: Chronos
        email: saul@popularowl.com
servers:
    - url: http://localhost:8001
      description: External HTTP agent (illustrative — your service)
    - url: http://localhost:3000/api/v1
      description: Chronos control plane (default dev)
...
security: []
paths:
...
    /api/v1/mcp-gateway/{agentId}/tools/invoke:
        post:
            tags:
                - mcp-gateway
            summary: Invoke an MCP tool through the Chronos gateway
            description: |
                Called by an external HTTP agent when it wants to use a
                registered MCP tool. The gateway:

                  1. Authenticates the agent's `mcpGatewayToken` (Bearer).
                  2. Parses the namespaced tool name `<server-slug>.<tool>`.
                  3. Enforces `Agent.allowedTools ∩ MCPServer.allowedTools`.
                  4. Opens (or reuses) a pooled MCP client to the resolved
                     `MCPServer` and calls `tools/call`.
                  5. Returns the MCP `CallToolResult`.

                The path is whitelisted from the platform's global API-key
                check — the per-agent gateway token IS the auth.

                A persistent audit table is planned for v1.7. v1.6 emits a
                structured `logger.info({ event: 'mcp.tool.invoke', … })`
                line per call.
            operationId: invokeMcpGatewayTool
            security:
                - mcpGatewayToken: []
            parameters:
                - in: path
                  name: agentId
                  required: true
                  schema:
                      type: string
                      format: uuid
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ToolInvokeRequest'
            responses:
                '200':
                    description: Tool returned a result.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ToolInvokeResponse'
                '400':
                    description: |
                        Malformed body (missing `tool`), unparseable
                        namespaced name, or the MCP server has no `url`
                        configured.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '401':
                    description: Missing / invalid Bearer or unknown agent. (Constant-time compare; existence is not leaked.)
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '403':
                    description: |
                        The agent is BUILT_IN or disabled, OR the requested
                        namespaced tool name is not in
                        `Agent.allowedTools ∩ MCPServer.allowedTools`.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Unknown MCP server slug.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '409':
                    description: MCP server is disabled.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '501':
                    description: MCP server uses `stdio` (reserved until v1.8).
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '502':
                    description: MCP transport error. The pooled client is evicted; the next call reconnects.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '503':
                    description: MCP server is unhealthy, OR the gateway itself is not enabled (`ENABLE_MCP_SERVERS=false`).
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /api/v1/mcp-gateway/{agentId}/tools:
        get:
            tags:
                - mcp-gateway
            summary: List the tools this agent is allowed to call
            description: |
                Returns the static intersection of `Agent.allowedTools` and
                each registered MCP server's `allowedTools`. Does NOT call
                `tools/list` on live MCP servers — for that, see
                `GET /api/v1/mcp-servers/{id}/tools` (operator surface).
            operationId: listAgentTools
            security:
                - mcpGatewayToken: []
            parameters:
                - in: path
                  name: agentId
                  required: true
                  schema:
                      type: string
                      format: uuid
            responses:
                '200':
                    description: Intersection of agent and server allowlists.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AllowedToolsResponse'
                '401':
                    description: Missing / invalid Bearer or unknown agent.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '403':
                    description: Agent is BUILT_IN or disabled.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
components:
    securitySchemes:
        mcpGatewayToken:
            type: http
            scheme: bearer
            bearerFormat: hex
            description: |
                Per-agent MCP gateway token. Stored on the `Agent` row as
                `mcpGatewayToken`. Rotatable from the Agent detail UI; rotation
                is single-step (no overlap window in v1.6).
    schemas:
...
        ToolInvokeRequest:
            type: object
            required:
                - tool
            properties:
                tool:
                    type: string
                    description: |
                        Namespaced tool name `<server-slug>.<tool-name>`,
                        e.g. `postgres.query`. Must be present in
                        `Agent.allowedTools` AND, unless empty,
                        `MCPServer.allowedTools`.
                    example: postgres.query
                params:
                    type: object
                    description: Arguments forwarded to the MCP tool as the `arguments` field of `tools/call`.
                    additionalProperties: true
                    example:
                        sql: select 1
                callId:
                    type: string
                    format: uuid
                    description: |
                        Optional correlation ID. Pass the
                        `x_chronos_call_id` from the inbound chat-completions
                        request so audit lines join up.
        ToolInvokeResponse:
            type: object
            required:
                - success
                - result
            properties:
                success:
                    type: boolean
                    const: true
                result:
                    type: object
                    description: |
                        Verbatim MCP `CallToolResult`. See the MCP
                        specification for the full shape — typically
                        `{ content: [{ type: 'text', text: '…' }] }` plus an
                        optional `isError` flag.
                    additionalProperties: true
        AllowedToolDescriptor:
            type: object
            required:
                - name
                - server
                - tool
            properties:
                name:
                    type: string
                    description: Namespaced name `<server>.<tool>`.
                    example: postgres.query
                server:
                    type: string
                    description: MCP server slug.
                    example: postgres
                tool:
                    type: string
                    description: Bare tool name on that server.
                    example: query
        AllowedToolsResponse:
            type: object
            required:
                - tools
            properties:
                tools:
                    type: array
                    items:
                        $ref: '#/components/schemas/AllowedToolDescriptor'
        Error:
            type: object
            required:
                - success
                - error
            properties:
                success:
                    type: boolean
                    const: false
                error:
                    type: string
...